server: add support for bool type in extraClaims
This commit add support for booleans in the app cap grant extraClaims.
This requires a few changes:

- add bool as a valid type in addClaimValue
- refactor flattenExtraClaims to keep original type information
- update test to test actual scalar values and not converted strings

Updates: #62
